refactor: split controller.go
- Split various functionality out of controller.go
- Add reporting.go file to controller package containing logic related to reporting deployments to the Artifact Metadata API
- Add workload package with encapsulated logic for resolving workload identity for pods.
